refactor #wip: get tests back to working state
