+++ x/common/testutil/README.md
@@ -1,4 +1,4 @@
 # Testutil Directory
 
-The x/testutil directory is not a cosmos-sdk module perse, but a collection of 
+The x/common/testutil directory is not a cosmos-sdk module perse, but a collection of 
 test utilities to make testing the other x/modules.
