+++ x/common/testutil/action/testcase.go
@@ -1,4 +1,4 @@
-package testutil
+package action
 
 import (
 	"testing"
