+++ x/perp/integration/action/vpool.go
@@ -4,9 +4,9 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/NibiruChain/nibiru/x/common/denoms"
+	"github.com/NibiruChain/nibiru/x/common/testutil/action"
 	"github.com/NibiruChain/nibiru/x/perp/keeper"
 	"github.com/NibiruChain/nibiru/x/perp/types"
-	"github.com/NibiruChain/nibiru/x/testutil"
 
 	"github.com/NibiruChain/nibiru/app"
 	"github.com/NibiruChain/nibiru/x/common/asset"
@@ -68,7 +68,7 @@ func CreateCustomVpool(
 	pair asset.Pair,
 	quoteAssetReserve, baseAssetReserve sdk.Dec,
 	vpoolConfig vpooltypes.VpoolConfig,
-) testutil.Action {
+) action.Action {
 	return CreateVPoolAction{
 		Pair:              pair,
 		QuoteAssetReserve: quoteAssetReserve,
